Optimization: using hash for trusted exe & arg check
Convert string comparison to 64-bit hash comparison, the former is less
efficient. With 128 items in the rbtree, 7 times of comparison will be
performed for every call of the hooked points.

Here murmur64 is used for string hash, and there's no collison among
18,734,347 pathes. hash is calculated only once when task is executed,
so the hash calculation won't bring any burden.

Signed-off-by: shenping.matt <shenping.matt@bytedance.com>